+++ kifi-backend/modules/common/app/com/keepit/common/oauth/OAuth2ConfigurationModule.scala
@@ -1,4 +1,4 @@
-package com.keepit.common.oauth2
+package com.keepit.common.oauth
 
 import java.net.URL
