+++ kifi-backend/modules/common/app/com/keepit/model/OAuth2TokenInfo.scala
@@ -1,6 +1,6 @@
 package com.keepit.model
 
-import com.keepit.common.oauth2.OAuth2AccessToken
+import com.keepit.common.oauth.OAuth2AccessToken
 import com.kifi.macros.json
 import play.api.libs.oauth.RequestToken
 import securesocial.core.{ OAuth1Info, OAuth2Info }
