+++ kifi-backend/modules/common/app/com/keepit/social/providers/FacebookProvider.scala
@@ -2,8 +2,8 @@ package com.keepit.social.providers
 
 import com.keepit.FortyTwoGlobal
 import com.keepit.common.logging.Logging
-import com.keepit.common.oauth2.FacebookOAuthProviderImpl
-import com.keepit.common.oauth2.adaptor.SecureSocialProviderHelper
+import com.keepit.common.oauth.FacebookOAuthProviderImpl
+import com.keepit.common.oauth.adaptor.SecureSocialProviderHelper
 import com.keepit.social.UserIdentityProvider
 
 import play.api.libs.ws.{ WSResponse }
