+++ kifi-backend/modules/shoebox/app/com/keepit/commanders/AuthCommander.scala
@@ -10,7 +10,7 @@ import com.keepit.common.db.{ ExternalId, Id }
 import com.keepit.common.db.slick.Database
 import com.keepit.common.healthcheck.AirbrakeNotifier
 import com.keepit.common.mail._
-import com.keepit.common.oauth2.{ ProviderIds, ProviderRegistry, OAuth2Configuration }
+import com.keepit.common.oauth.{ ProviderIds, ProviderRegistry, OAuth2Configuration }
 import com.keepit.common.performance.timing
 import com.keepit.common.store.{ ImageCropAttributes, S3ImageStore }
 import com.keepit.common.time.Clock
